Изменение интерфейса.Получение ответа API quotes

+++ HikikomoriWEB/Controllers/ContentController.cs
@@ -3,6 +3,7 @@ using HikikomoriWEB.Services.Interfaces;
 using HikikomoriWEB.Domain.Entity;
 using HikikomoriWEB.Domain.ViewModels;
 using System.Threading.Tasks;
+using HikikomoriWEB.Services.HelperMethods;
 
 namespace HikikomoriWEB.Controllers
 {
@@ -22,11 +23,10 @@ namespace HikikomoriWEB.Controllers
         {
             var rateResponse = await _rateService.GetFilms();
             var rememberResponse = await _rememberService.GetFilms();
-            if(rateResponse.StatusCode == Domain.Enum.StatusCode.ServerError || rememberResponse.StatusCode == Domain.Enum.StatusCode.ServerError)
+            if(ControllerAssistant.ErrorCheck<RateContent, RememberContent>(rateResponse, rememberResponse))
             {
                 return RedirectToAction("Error");
             }
-
             return View(new ContentListViewModel(rateResponse.Data, rememberResponse.Data));
         }
 
@@ -34,11 +34,10 @@ namespace HikikomoriWEB.Controllers
         {
             var rateResponse = await _rateService.GetBooks();
             var rememberResponse = await _rememberService.GetBooks();
-            if (rateResponse.StatusCode == Domain.Enum.StatusCode.ServerError || rememberResponse.StatusCode == Domain.Enum.StatusCode.ServerError)
+            if (ControllerAssistant.ErrorCheck<RateContent, RememberContent>(rateResponse, rememberResponse))
             {
                 return RedirectToAction("Error");
             }
-
             return View(new ContentListViewModel(rateResponse.Data, rememberResponse.Data));
         }
 
@@ -46,11 +45,10 @@ namespace HikikomoriWEB.Controllers
         {
             var rateResponse = await _rateService.GetGames();
             var rememberResponse = await _rememberService.GetGames();
-            if (rateResponse.StatusCode == Domain.Enum.StatusCode.ServerError || rememberResponse.StatusCode == Domain.Enum.StatusCode.ServerError)
+            if (ControllerAssistant.ErrorCheck<RateContent, RememberContent>(rateResponse, rememberResponse))
             {
                 return RedirectToAction("Error");
             }
-
             return View(new ContentListViewModel(rateResponse.Data, rememberResponse.Data));
         }
 
@@ -58,11 +56,10 @@ namespace HikikomoriWEB.Controllers
         {
             var rateResponse = await _rateService.GetSerials();
             var rememberResponse = await _rememberService.GetSerials();
-            if (rateResponse.StatusCode == Domain.Enum.StatusCode.ServerError || rememberResponse.StatusCode == Domain.Enum.StatusCode.ServerError)
+            if (ControllerAssistant.ErrorCheck<RateContent, RememberContent>(rateResponse, rememberResponse))
             {
                 return RedirectToAction("Error");
             }
-
             return View(new ContentListViewModel(rateResponse.Data, rememberResponse.Data));
         }
 
@@ -70,27 +67,29 @@ namespace HikikomoriWEB.Controllers
         {
             var rateResponse = await _rateService.GetCartoons();
             var rememberResponse = await _rememberService.GetCartoons();
-            if (rateResponse.StatusCode == Domain.Enum.StatusCode.ServerError || rememberResponse.StatusCode == Domain.Enum.StatusCode.ServerError)
+            if (ControllerAssistant.ErrorCheck<RateContent, RememberContent>(rateResponse, rememberResponse))
             {
                 return RedirectToAction("Error");
             }
-
             return View(new ContentListViewModel(rateResponse.Data, rememberResponse.Data));
         }
         #endregion
 
-        public async Task<IActionResult> RemoveActon(int Id, string tableClass)
+        public async Task<IActionResult> RemoveAction(int Id, string tableClass)
         {
-            switch (tableClass)
+            if(Id != 0 && tableClass != null)
             {
-                case "table-list-rate" :
-                    await _rateService.DeleteContent(Id);
-                    break;
-                case "table-list-remember":
-                    await _rememberService.DeleteContent(Id);
-                    break;
+                switch (tableClass)
+                {
+                    case "table-list-rate":
+                        await _rateService.DeleteContent(Id);
+                        break;
+                    case "table-list-remember":
+                        await _rememberService.DeleteContent(Id);
+                        break;
+                }
             }
-            return RedirectToAction("Index"); //изменить
+            return NoContent();
         }
     }
 }
